pyproject.toml
[project]
name = "inventory-management-system-api"
description = "Python microservice providing an API for the Inventory Management System."
readme = "README.md"
requires-python = ">=3.13"
version = "3.0.0"

dependencies = [
    "cryptography",
    "fastapi[all]",
    "PyJWT",
    "pymongo",
    "requests",
    "uvicorn"
]

[project.urls]
"Repository" = "https://github.com/ral-facilities/inventory-management-system-api"

[project.scripts]
"ims" = "inventory_management_system_api.cli.main:main"

[project.optional-dependencies]
code-analysis = [
    "black==25.9.0",
    "pylint==3.3.9",
    "inventory-management-system-api[test]"
]

test = [
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "pytest-cov==7.0.0",
    "pytest-env==1.2.0"
]

scripts = [
    "faker==37.11.0"
]

dev = [
    "inventory-management-system-api[code-analysis]",
    "inventory-management-system-api[test]",
    "inventory-management-system-api[scripts]"
]

[tool.coverage.run]
omit = [
    # Exclude test code from coverage measurement
    "test/*",
    # Exclude migration scripts from coverage measurement
    "inventory_management_system_api/migrations/scripts/*",
    # Exclude routers from coverage measurement as they are tested in the e2e tests
    "inventory_management_system_api/routers/v1/*"
]

[tool.setuptools.packages.find]
include  = ["inventory_management_system_api", "inventory_management_system_api*"]
